U-Net/requirements.txt
matplotlib
numpy
Pillow
torch
torchvision
tensorboard
future
imgaug
